Tophat: Include examples through a webpack loader
This makes the examples part of the build pipeline instead of something
that gets serialised into the page. This makes individual pages smaller
at the expense of a larger bundle (which is shared amongst pages).

This means that you can now edit an example then refresh the page in
your browser to see changes instead of having to fully restart the
server. This also opens the door to follow-up PRs to enable
hot-reloading.