Unify playground and tophat servers
Run the playground server with `yarn run dev` then visit /examples to
see all the examples that were present as part of the tophat server

This means that examples are now hotreloadable - modifications to README
files and components will be reflected in your browser without needing
to save.

Rewrite most of the tophat app into a more app-like folder structure
that lives in /playground/app.